chatbot-tvts-KMS/requirements.txt

flask==3.0.0
flask-cors==4.0.0

chromadb==0.5.5
psycopg2-binary==2.9.9

openai>=1.3.7

pandas==2.1.4
numpy==1.26.2
python-Levenshtein==0.23.0

requests==2.31.0

beautifulsoup4==4.12.2

python-dotenv==1.0.0

panel==1.6
param==2.0.1
markdown==3.5.1

python-dateutil==2.8.2
pytz==2023.3.post1

tiktoken

